requirements.txt
Django>=4.2
psycopg2-binary
celery
redis
pytest
pytest-django
djangorestframework

